requirements.txt
aioamqp==0.12.0
amqp==2.4.2
asgiref==2.3.2
async-timeout==3.0.1
attrs==19.1.0
autobahn==19.2.1
Automat==0.7.0
billiard==3.5.0.5
bitcart==0.1
celery==4.2.1
certifi==2018.11.29
channels==2.1.7
channels-rabbitmq==0.0.11
chardet==3.0.4
constantly==15.1.0
coreapi==2.3.3
coreschema==0.0.4
daphne==2.2.5
Django==2.1.7
django-embed-video==1.2.0
django-filter==2.1.0
djangorestframework==3.9.2
hyperlink==18.0.0
idna==2.8
incremental==17.5.0
itypes==1.1.0
Jinja2==2.10
kombu==4.4.0
Markdown==3.0.1
MarkupSafe==1.1.1
msgpack==0.6.1
mysqlclient==1.4.2.post1
Pillow==5.4.1
Pygments==2.3.1
PyHamcrest==1.9.0
python-decouple==3.1
pytz==2018.9
requests==2.21.0
six==1.12.0
sorl-thumbnail==12.5.0
Twisted==18.9.0
txaio==18.8.1
uritemplate==3.0.0
urllib3==1.24.1
vine==1.2.0
whitenoise==4.1.2
zope.interface==4.6.0
